fix(graph): bind source production to retrieval surface
